Update mail path

+++ infra/deploy/deploy.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-# Migrates DB
 cd app
 git pull origin master
 
@@ -9,19 +8,26 @@ cd backend
 export NVM_DIR="/home/ubuntu/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
 
-npx prisma migrate deploy
+if npx prisma migrate deploy; then
+    cd ~
 
-cd ~
+    # Login to Docker Hub
+    echo "$DOCKER_HUB_PASSWORD" | docker login --username "$DOCKER_HUB_USERNAME" --password-stdin
 
-# Login to Docker Hub
-echo "$DOCKER_HUB_PASSWORD" | docker login --username "$DOCKER_HUB_USERNAME" --password-stdin
+    # Pull the latest images
+    docker pull vcarp/photoscorer:latest
+    docker pull vcarp/photoscorer-nginx:latest
 
-# Pull the latest images
-docker pull vcarp/photoscorer:latest
-docker pull vcarp/photoscorer-nginx:latest
+    # Use Docker Compose to restart the services
+    docker-compose down
+    docker-compose up -d
 
-# Use Docker Compose to restart the services
-docker-compose down
-docker-compose up -d
+    echo "Update complete."
+else
+    # Migration failed, handle the error as needed
+    # You can log the error, send notifications, or take other actions
+    echo "Database migration failed. Check logs for details."
 
-echo "Update complete."
+    # Optionally, you can exit the script to prevent further execution
+    exit 1
+fi
